+++ RELEASE_NOTES.md
@@ -28,7 +28,7 @@ If you upgrade from an older version to `3.0` and have many breaking changes, up
 ### Non-breaking changes:
 
 * `KeyboardInputViewController` `deviceOrientation` has been converted to a general `UIInputViewController` extension.
-* `KeyboardInputViewController` `setupNextKeyboardButton` has been converted to a general `UIInputViewController` extension.
+* `setupNextKeyboardButton` has been converted to a general `UIInputViewController` extension.
 
 
 ### Breaking changes:
@@ -39,6 +39,7 @@ If you upgrade from an older version to `3.0` and have many breaking changes, up
 * `KeyboardActionRows.from` has been changed to an initializer.
 * `KeyboardImageActions` has been converted to a `KeyboardActionRow+Images` extension initializer.
 * `KeyboardInputViewController` has a new `keyboardContext`.
+* `KeyboardInputViewController` `canChangeKeyboardType` has been moved to `KeyboardType`.
 * `KeyboardInputViewController` `keyboardType` has been moved to the context.
 * `KeyboardToolbar` has been renamed to `KeyboardToolbarView`. 
 * The `shouldChangeToAlphabeticLowercase` has been replaced with the automatic switching mentioned above.